Altered tasktree.Cmd to inherit from the vanilla object type.
Refactored spool method of serialize to accept a new argument 'user' and
to append on each non-default argument to the command line invocation rather
than the previous if/else preformatted text.

This means a list is passed to subprocess rather than a single string. At this
point since the call is made within a try clause there is very little chance
of a return code being returned in the event of an exception. Printing the return
code in the except case has therefore been removed.